Implement useReducer w/ useContext for AppState management
Added AppContext, which currently keeps track of the current page. This
is so the header can render the appropriate version of an icon, depending
on if the page is active or not. The main pages now take a dispatch function
in their props, they can fire off a dispatch with the "changePage" type and
a payload of the Page enum to update the AppContext.