Implement react-easy-crop into addPost modal
After selecting a file, the user can now crop it into a square image.
After they advance through the modal they are given the option to share
the image, which then uploads a cropped version of the image to
firebase storage, and creates a post in firestore that references the file
storage.