Expand addPost modal on PostStage.ShareImage
When you make it to the share stage of add post, the modal expands to
fit a right aligned bar to add your comment. Currently the comment
functionality is not implemented. This was a fair amount of work to get the
crop image to be the correct size, and to have the modal scale vertically.
I am glad this is working.